+++ src/plugins/sdcard/src/android/SDcard.java
@@ -554,8 +554,13 @@ public class SDcard extends CordovaPlugin {
 
             try {
               DocumentFile file = DocumentFile.fromTreeUri(context, fileUri);
+              // If only case change, OS adds '(<number>)' as suffix, to avoid that we need to rename to a temporary name first
+              if (newFile.equalsIgnoreCase(file.getName())) {
+                file.renameTo(newFile + "_temp");
+              }
 
               if (file.renameTo(newFile)) {
+                String name = file.getName();
                 docId = DocumentsContract.getDocumentId(file.getUri());
                 callback.success(srcUri + SEPARATOR + docId);
               } else {
@@ -852,6 +857,7 @@ public class SDcard extends CordovaPlugin {
               result.put("isDirectory", file.isDirectory());
               result.put("isVirtual", file.isVirtual());
               result.put("lastModified", file.lastModified());
+              result.put("url", file.getUri().toString());
 
               callback.success(result);
             } catch (Exception e) {