+++ src/plugins/system/android/com/foxdebug/system/System.java
@@ -28,7 +28,7 @@ import androidx.core.content.FileProvider;
 import androidx.core.content.pm.ShortcutInfoCompat;
 import androidx.core.content.pm.ShortcutManagerCompat;
 import androidx.core.graphics.drawable.IconCompat;
-import com.foxdebug.system.BrowserDialog;
+import com.foxdebug.system.Ui.Theme;
 import java.io.File;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
@@ -49,15 +49,13 @@ import org.json.JSONObject;
 
 public class System extends CordovaPlugin {
 
-  // private WebviewActivity webviewActivity;
-  private BrowserDialog browserDialog;
   private CallbackContext requestPermissionCallback;
   private Activity activity;
   private Context context;
   private int REQ_PERMISSIONS = 1;
   private int REQ_PERMISSION = 2;
-  private int themeColor = 0xFF000000;
-  private String themeType = "dark";
+  private int systemBarColor = 0xFF000000;
+  private Theme theme;
   private CallbackContext intentHandler;
   private CordovaWebView webView;
 
@@ -117,23 +115,7 @@ public class System extends CordovaPlugin {
           .runOnUiThread(
             new Runnable() {
               public void run() {
-                setUiTheme(arg1, arg2, callbackContext);
-              }
-            }
-          );
-        return true;
-      case "in-app-browser":
-        this.cordova.getActivity()
-          .runOnUiThread(
-            new Runnable() {
-              public void run() {
-                inAppBrowser(
-                  arg1,
-                  arg2,
-                  args.optBoolean(2),
-                  args.optBoolean(3),
-                  callbackContext
-                );
+                setUiTheme(arg1, args.optJSONObject(1), callbackContext);
               }
             }
           );
@@ -573,52 +555,6 @@ public class System extends CordovaPlugin {
     callback.success("Launched " + appId);
   }
 
-  private void inAppBrowser(
-    String url,
-    String title,
-    boolean showButtons,
-    Boolean disableCache,
-    CallbackContext callback
-  ) {
-    BrowserDialog browserDialog = new BrowserDialog(
-      this,
-      themeColor,
-      themeType,
-      showButtons,
-      disableCache,
-      callback
-    );
-    browserDialog.show(url, title);
-    this.browserDialog = browserDialog;
-  }
-
-  public void onActivityResult(int requestCode, int resultCode, Intent intent) {
-    if (
-      requestCode != browserDialog.FILE_SELECT_CODE ||
-      browserDialog.webViewFilePathCallback == null
-    ) {
-      super.onActivityResult(requestCode, resultCode, intent);
-      return;
-    }
-    ArrayList<Uri> uris = new ArrayList<Uri>();
-    ClipData clipData = intent.getClipData();
-    if (clipData != null) {
-      for (int i = 0; i < clipData.getItemCount(); i++) {
-        ClipData.Item item = clipData.getItemAt(i);
-        Uri uri = item.getUri();
-        uris.add(uri);
-      }
-    } else {
-      Uri uri = intent.getData();
-      uris.add(uri);
-    }
-
-    browserDialog.webViewFilePathCallback.onReceiveValue(
-      uris.toArray(new Uri[uris.size()])
-    );
-    browserDialog.webViewFilePathCallback = null;
-  }
-
   private void addShortcut(
     String id,
     String label,
@@ -706,15 +642,15 @@ public class System extends CordovaPlugin {
   }
 
   private void setUiTheme(
-    final String color,
-    final String type,
+    final String systemBarColor,
+    final JSONObject theme,
     final CallbackContext callback
   ) {
+    this.systemBarColor = Color.parseColor(systemBarColor);
+    this.theme = new Theme(theme);
+
     if (Build.VERSION.SDK_INT >= 21) {
-      final int bgColor = Color.parseColor(color);
       final Window window = activity.getWindow();
-      themeColor = bgColor;
-      themeType = type.toLowerCase();
       // Method and constants not available on all SDKs but we want to be able to compile this code with any SDK
       window.clearFlags(0x04000000); // SDK 19: WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS);
       window.addFlags(0x80000000); // SDK 21: WindowManager.LayoutParams.FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS);
@@ -724,12 +660,12 @@ public class System extends CordovaPlugin {
         window
           .getClass()
           .getMethod("setNavigationBarColor", int.class)
-          .invoke(window, themeColor);
+          .invoke(window, this.systemBarColor);
 
         window
           .getClass()
           .getMethod("setStatusBarColor", int.class)
-          .invoke(window, themeColor);
+          .invoke(window, this.systemBarColor);
 
         setStatusBarStyle(window);
         setNavigationBarStyle(window);
@@ -746,6 +682,7 @@ public class System extends CordovaPlugin {
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
       View decorView = window.getDecorView();
       int uiOptions = decorView.getSystemUiVisibility();
+      String themeType = theme.getType();
 
       if (themeType.equals("light")) {
         decorView.setSystemUiVisibility(
@@ -763,6 +700,7 @@ public class System extends CordovaPlugin {
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
       View decorView = window.getDecorView();
       int uiOptions = decorView.getSystemUiVisibility();
+      String themeType = theme.getType();
 
       // 0x80000000 FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS
       // 0x00000010 SYSTEM_UI_FLAG_LIGHT_NAVIGATION_BAR